- Replaced FCK editor with TinyMCE
- Autotranslation of XUL templates
- --standalone command line flag
- removal of gtk :(



git-svn-id: file:///home/jwt/exe/trunk@1314 38b22f21-9aea-0310-abfc-843a9883df58

+++ exe/xului/outlinepane.py
@@ -62,10 +62,10 @@ class OutlinePane(Renderable):
             elif (nodeId != package.root.id and 
                   request.args["action"][0] == "deleteNode"):
                 node = package.findNode(nodeId)
-                if node is not None:
+                if node is not None and node is not self.package.root:
                     node.delete()
                     if node.isAncestorOf(package.currentNode):
-                        package.currentNode = package.root
+                        package.currentNode = node.parent
                 else:
                     log.error("deleteNode cannot locate "+nodeId)
